Fixed lint stealing focus in OS X
Summary: When the Android linter would run in OS X, a Java icon would briefly appear in the dock and steal focus. Now it's forced to run in "headless" mode, so no GUI elements appear for lint execution.

Test Plan: Run `arc lint` on an Android project and watch for the dock icon.

Reviewers: mbishop

Reviewed By: mbishop

Maniphest Tasks: T3591

Differential Revision: https://devportal.imobile3.com/D3572