Fixed Android unit testing issue with APK installation
Summary: APKs were occasionally not getting installed. Apparently `exec_manual()` adds quotes around the `%s` strings, so it was quoting an entire switch like this: `adb '-s 123456789' install ...`. Now, it does this: `adb -s '123456789' install ...`. Also adjusted the login to determine when APK installation fails and the login that waits for a device if no device is plugged in.

Test Plan:
# Uninstall all app and test packages from a device for a project.
# Run `arc unit` on that Android project with a device plugged in.
# Uninstall packages again
# Run `arc unit` again with a device unplugged. Plug it in when prompted.

Reviewers: mbishop, mportune

Reviewed By: mbishop

Differential Revision: https://devportal.imobile3.com/D5511