Linter for improper BigDecimal usage
Summary: It's really easy to use BigDecimal in an unsafe way. Like, REALLY easy. Hopefully this will help.

Test Plan: Run linter on any POS project; I'm sure it'll pick out a few disappointments. Note that I tried adding unit tests for the new linter, but because it's a class linter instead of a Java source linter it depends on having a compiled project and the tests would not work.

Reviewers: jstarke, mbishop, mportune

Reviewed By: mportune

Subscribers: mbishop

Differential Revision: https://devportal.imobile3.com/D12018