Updated Android filetypes
Summary: Added `.gradle` and `.properties` as file extensions that trigger the Android linter to run. If you tried to lint a diff that only updated files of these types, it would always run the linter as if it were an Eclipse project.

Test Plan: `arc patch D12253` and then `arc lint`

Reviewers: mportune, mbishop

Reviewed By: mbishop

Subscribers: mbishop

Differential Revision: https://devportal.imobile3.com/D12258