Cassowary - iOS Static Analysis Linter Improvements
Summary:
The existing codebase only supports reading the first compiler error or warning per file. This revision updates the linter to read each line generated for the file and produce all errors and warnings.

Additionally this revision updates the regex used to produce the error or warning message to accomodate the full message. Previously, messages could be cut off by special characters, resulting in confusingly truncated errors.

Test Plan: Create a project with multiple static analysis or compiler warnings and errors in the same file. Run "arc lint" on the file and observe that multiple warnings and errors are now produced, with their full message intact.

Reviewers: mbishop, ablackwood

Reviewed By: ablackwood

Subscribers: mbishop

Differential Revision: https://devportal.imobile3.com/D12295