Fix for Gradle linter
Summary: The Gradle linter returns a non-zero exit code if it finds a lint error. Since usually it will only find warnings, the exit code looks like a reliable way to determine its success or failure. Now, the existnace of the output files is used to determine success or failure instead.

Test Plan: Introduce a lint error into an Android project and run `arc lint` on it.

Reviewers: mbishop, mportune

Reviewed By: mportune

Subscribers: mbishop

Differential Revision: https://devportal.imobile3.com/D13041