Forcing reporter type of 'JSON' in ArcanistSwiftLinter
Summary: The swift linter in Cassowary was set up to respect the reporter type in the .yml file. This is fine but it also requires that said reporter type be set to JSON. For our purposes, that reporter type needs to be set to Xcode so that lint warnings can be presented while development is underway. By hardcoding this to always use JSON, we ensure the arc linter will always work and the local reporter type (which can only be one or the other) can be set to what the developer wants for his/her purposes.

Test Plan: Confirmed swiftlint works in Xcode but also does not break arc diff.

Reviewers: tlahtinen, jdunnivant, mbishop

Reviewed By: jdunnivant, mbishop

Differential Revision: https://phab.imobile3.local/D13210