Resolve unit test loop where a deleted folder prevents the unit test engine from moving up the file tree
Summary: A deleted folder was causing the unit test engine to continue in a loop since the file path of the deleted folder didn't exist. The change is to keep going up the file tree until the file path exists.

Test Plan: Validate unit tests run successfully when there is a deleted folder included in a commit. D13214 encountered this issue and will be used to test with.

Reviewers: mbishop, jdunnivant

Reviewed By: mbishop

JIRA Issues: ARCH-78

Differential Revision: https://phab.imobile3.local/D13298